refactor(dictation): 抽出 dictation_streaming 子模块
把流式润色与文本收尾相关函数从 coordinator/dictation.rs 机械式搬到
同级 coordinator/dictation_streaming.rs（行为保持）：run_streaming_polish、
drain/flush streaming insert、finalize_polished_text、streaming_insert_eligible、
default_done_message、append_typed_prefix，以及 STREAMING_INSERT_FLUSH_INTERVAL。
coordinator.rs glob 重导出新模块，dictation.rs 测试改走 crate::coordinator:: 路径。
